requirements.txt
sleekxmpp=1.3.1
dnspython3=1.12.0
pyasn1=0.1.9
pyasn1-modules=0.0.8
pyyaml=3.10
pydle=0.8.2